Fix uint8_t overflow in rejection sampling threshold
When num_channels-1 divides 256 evenly (e.g. 3-channel AU433/EU433 domains),
256 % range == 0 and threshold = 256, which overflows uint8_t to 0. This
caused every random byte to be rejected, burning all 16 retries per hop
and repeatedly regenerating the XOF buffer.

Fix: use uint16_t for threshold so 256 is representable. When threshold is
256, no values are rejected (correct: no modulo bias when range divides 256).

Added test for 3-channel domain edge case. 41/41 tests pass.

Found by Codex (gpt-5.3-codex) code review.

+++ src/lib/MurmurEncrypt/murmur.c
@@ -153,7 +153,9 @@ void murmur_fhss_fill_sequence(const uint8_t fhss_key[16],
      * Rejection sampling eliminates modulo bias: we discard values >= threshold
      * where threshold is the largest multiple of (num_channels-1) that fits in 256. */
     uint8_t range = num_channels - 1;
-    uint8_t threshold = 256 - (256 % range);
+    /* Use uint16_t: when range divides 256, 256 % range == 0 and threshold == 256,
+     * which overflows uint8_t to 0 (every sample would be rejected). */
+    uint16_t threshold = 256 - (256 % range);
     uint16_t buf_idx = 0;
 
     for (uint16_t i = 0; i < seq_len; i++) {
